Cannot cache build image tasks
We can't reliably assume that the image/tag actually corresponds
to the same inputs when switch between changes we already built before.

As gradle happy places the generated digest.json file and ignores that
the tagged image hasn't actually been changed.

This is because gradle isn't actually tracking the generated image but
rather its approximate digest. So the tag can point to the previous
build when switching between two sets of changes for which gradle has
a cached output.